+++ sdk/communication/communication-chat/src/chatClient.ts
@@ -139,6 +139,7 @@ export class ChatClient {
               mapToChatParticipantRestModel(participant),
             ),
             metadata: options.metadata,
+            retentionPolicy: options.retentionPolicy,
           },
           updatedRestModelOptions,
         );