+++ sdk/communication/communication-chat/test/public/chatClient.spec.ts
@@ -3,7 +3,7 @@
 
 import { Recorder, isLiveMode } from "@azure-tools/test-recorder";
 import { assert } from "chai";
-import { ChatClient, ChatThreadClient } from "../../src";
+import { ChatClient, ChatThreadClient, ThreadCreationDateRetentionPolicy } from "../../src";
 import { createChatClient, createRecorder, createTestUser } from "./utils/recordedClient";
 import { isNode } from "@azure/core-util";
 import sinon from "sinon";
@@ -45,6 +45,10 @@ describe("ChatClient", function () {
       testUser2 = (await createTestUser(recorder)).user;
 
       const request = { topic: "test topic" };
+      const retentionPolicy: ThreadCreationDateRetentionPolicy = {
+        kind: "threadCreationDate",
+        deleteThreadAfterDays: 90,
+      };
       const options = {
         participants: [
           {
@@ -58,6 +62,7 @@ describe("ChatClient", function () {
           threadType: "primary",
           secondaryThread: "test-id",
         },
+        retentionPolicy: retentionPolicy,
       };
 
       const chatThreadResult = await chatClient.createChatThread(request, options);
@@ -70,6 +75,7 @@ describe("ChatClient", function () {
       assert.isDefined(chatThread?.id);
       assert.equal(chatThread?.topic, request.topic);
       assert.deepEqual(chatThread?.metadata, options.metadata);
+      assert.deepEqual(chatThread?.retentionPolicy, retentionPolicy);
     }).timeout(8000);
 
     it("successfully retrieves a thread client", async function () {